// types/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};
// ...
export type AcfFieldGroup = {
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
};
// ...
export type Page_Pagepanels_PagePanels_WorkSlider = AcfFieldGroup & {
  __typename?: 'Page_Pagepanels_PagePanels_WorkSlider';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  slider?: Maybe<Array<Maybe<Page_Pagepanels_PagePanels_WorkSlider_Slider>>>;
};

/** Field Group */
export type Page_Pagepanels_PagePanels_WorkSlider_Slider = AcfFieldGroup & {
  __typename?: 'Page_Pagepanels_PagePanels_WorkSlider_slider';
  featuredProject?: Maybe<Page_Pagepanels_PagePanels_WorkSlider_Slider_FeaturedProject>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
};

export type Page_Pagepanels_PagePanels_WorkSlider_Slider_FeaturedProject = Project;
// ...
export type Project = ContentNode & DatabaseIdentifier & MenuItemLinkable & Node & NodeWithExcerpt & NodeWithRevisions & NodeWithTemplate & NodeWithTitle & UniformResourceIdentifiable & {
  __typename?: 'Project';
  /** Connection between the ContentNode type and the ContentType type */
  contentType?: Maybe<ContentNodeToContentTypeConnectionEdge>;
  /** The name of the Content Type the node belongs to */
  contentTypeName: Scalars['String'];
  /** The unique resource identifier path */
  databaseId: Scalars['Int'];
  /** Post publishing date. */
  date?: Maybe<Scalars['String']>;
  /** The publishing date set in GMT. */
  dateGmt?: Maybe<Scalars['String']>;
  /** The desired slug of the post */
  desiredSlug?: Maybe<Scalars['String']>;
  /** If a user has edited the node within the past 15 seconds, this will return the user that last edited. Null if the edit lock doesn&#039;t exist or is greater than 15 seconds */
  editingLockedBy?: Maybe<ContentNodeToEditLockConnectionEdge>;
  /** The RSS enclosure for the object */
  enclosure?: Maybe<Scalars['String']>;
  /** Connection between the ContentNode type and the EnqueuedScript type */
  enqueuedScripts?: Maybe<ContentNodeToEnqueuedScriptConnection>;
  /** Connection between the ContentNode type and the EnqueuedStylesheet type */
  enqueuedStylesheets?: Maybe<ContentNodeToEnqueuedStylesheetConnection>;
  /** The excerpt of the post. */
  excerpt?: Maybe<Scalars['String']>;
  /** The global unique identifier for this post. This currently matches the value stored in WP_Post-&gt;guid and the guid column in the &quot;post_objects&quot; database table. */
  guid?: Maybe<Scalars['String']>;
  /** The globally unique identifier of the work object. */
  id: Scalars['ID'];
  /** Whether the node is a Content Node */
  isContentNode: Scalars['Boolean'];
  /** Whether the object is a node in the preview state */
  isPreview?: Maybe<Scalars['Boolean']>;
  /** Whether the object is restricted from the current viewer */
  isRestricted?: Maybe<Scalars['Boolean']>;
  /** True if the node is a revision of another node */
  isRevision?: Maybe<Scalars['Boolean']>;
  /** Whether the node is a Term */
  isTermNode: Scalars['Boolean'];
  /** The user that most recently edited the node */
  lastEditedBy?: Maybe<ContentNodeToEditLastConnectionEdge>;
  /** The permalink of the post */
  link?: Maybe<Scalars['String']>;
  /** The local modified time for a post. If a post was recently updated the modified field will change to match the corresponding time. */
  modified?: Maybe<Scalars['String']>;
  /** The GMT modified time for a post. If a post was recently updated the modified field will change to match the corresponding time in GMT. */
  modifiedGmt?: Maybe<Scalars['String']>;
  /** Next project */
  next?: Maybe<Project>;
  /** Added to the GraphQL Schema because the ACF Field Group &quot;Page Panels&quot; was set to Show in GraphQL. */
  pagePanels?: Maybe<Project_Pagepanels>;
  /** Connection between the project type and the project type */
  preview?: Maybe<ProjectToPreviewConnectionEdge>;
  /** The database id of the preview node */
  previewRevisionDatabaseId?: Maybe<Scalars['Int']>;
  /** Whether the object is a node in the preview state */
  previewRevisionId?: Maybe<Scalars['ID']>;
  /** Previous project */
  previous?: Maybe<Project>;
  /** Added to the GraphQL Schema because the ACF Field Group &quot;Project - Brand Colour&quot; was set to Show in GraphQL. */
  projectBrandColour?: Maybe<Project_Projectbrandcolour>;
  /**
   * The id field matches the WP_Post-&gt;ID field.
   * @deprecated Deprecated in favor of the databaseId field
   */
  projectId: Scalars['Int'];
  /** If the current node is a revision, this field exposes the node this is a revision of. Returns null if the node is not a revision of another node. */
  revisionOf?: Maybe<NodeWithRevisionsToContentNodeConnectionEdge>;
  /** Connection between the project type and the project type */
  revisions?: Maybe<ProjectToRevisionConnection>;
  /** The Yoast SEO data of the ContentNode */
  seo?: Maybe<PostTypeSeo>;
  /** The uri slug for the post. This is equivalent to the WP_Post-&gt;post_name field and the post_name column in the database for the &quot;post_objects&quot; table. */
  slug?: Maybe<Scalars['String']>;
  /** The current status of the object */
  status?: Maybe<Scalars['String']>;
  /** The template assigned to the node */
  template?: Maybe<ContentTemplate>;
  /** The title of the post. This is currently just the raw title. An amendment to support rendered title needs to be made. */
  title?: Maybe<Scalars['String']>;
  /** The unique resource identifier path */
  uri?: Maybe<Scalars['String']>;
};
// ...
export type Project_Pagepanels_PagePanels_WorkSlider = AcfFieldGroup & {
  __typename?: 'Project_Pagepanels_PagePanels_WorkSlider';
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
  slider?: Maybe<Array<Maybe<Project_Pagepanels_PagePanels_WorkSlider_Slider>>>;
};

/** Field Group */
export type Project_Pagepanels_PagePanels_WorkSlider_Slider = AcfFieldGroup & {
  __typename?: 'Project_Pagepanels_PagePanels_WorkSlider_slider';
  featuredProject?: Maybe<Project_Pagepanels_PagePanels_WorkSlider_Slider_FeaturedProject>;
  /** The name of the ACF Field Group */
  fieldGroupName?: Maybe<Scalars['String']>;
};

export type Project_Pagepanels_PagePanels_WorkSlider_Slider_FeaturedProject = Project;